+++ src/Geopilot.Api/ContextExtensions.cs
@@ -125,7 +125,7 @@ internal static class ContextExtensions
         var longMax = new Faker().Random.Double(extentCh[2] - longDiffHalf, extentCh[2]);
         var latMax = new Faker().Random.Double(extentCh[3] - latDiffHalf, extentCh[3]);
 
-        return GeometryFactory.Default.CreatePolygon(new Coordinate[]
+        return GeometryFactory.Default.CreatePolygon(new NetTopologySuite.Geometries.Coordinate[]
         {
             new (longMin, latMin),
             new (longMax, latMin),
@@ -139,7 +139,6 @@ internal static class ContextExtensions
     {
         var knownFileFormats = new string[] { ".xtf", ".gpkg", ".*", ".itf", ".xml", ".zip", ".csv" };
         var mandateFaker = new Faker<Mandate>()
-            .StrictMode(true)
             .RuleFor(o => o.Id, f => 0)
             .RuleFor(o => o.Name, f => f.Commerce.ProductName())
             .RuleFor(o => o.FileTypes, f => f.PickRandom(knownFileFormats, 4).Distinct().ToArray())
@@ -168,7 +167,7 @@ internal static class ContextExtensions
             .RuleFor(d => d.Date, f => f.Date.Past().ToUniversalTime())
             .RuleFor(d => d.Mandate, f => f.PickRandom(deliveryContracts))
             .RuleFor(d => d.DeclaringUser, (f, d) => f.PickRandom(
-                d.Mandate.Organisations
+                d.Mandate!.Organisations
                 .SelectMany(o => o.Users)
                 .ToList()))
             .RuleFor(d => d.Assets, _ => new List<Asset>())