+++ src/Geopilot.Frontend/src/AppInterfaces.ts
@@ -1,4 +1,10 @@
-import { User } from "./auth/AuthInterfaces";
+export type Language = "de" | "fr" | "it" | "en";
+
+export interface TranslationFunction {
+  (key: string): string;
+}
+
+export type ModalContentType = "markdown" | "raw";
 
 export interface ClientSettings {
   authCache: {
@@ -19,16 +25,34 @@ export interface ClientSettings {
   theme: object;
 }
 
-export type Language = "de" | "fr" | "it" | "en";
-
-export interface TranslationFunction {
-  (key: string): string;
+export interface Validation {
+  allowedFileExtensions: string[];
 }
 
-export type ModalContentType = "markdown" | "raw";
+export interface ErrorResponse {
+  status: string;
+  detail: string;
+}
+
+export interface Coordinate {
+  x: number | null;
+  y: number | null;
+}
 
 export interface Mandate {
+  id: number;
   name: string;
+  fileTypes: string[];
+  coordinates: Coordinate[];
+  organisations: Organisation[] | number[];
+  deliveries: Delivery[];
+}
+
+export interface Organisation {
+  id: number;
+  name: string;
+  mandates: Mandate[] | number[];
+  users: User[] | number[];
 }
 
 export interface Delivery {
@@ -37,4 +61,13 @@ export interface Delivery {
   declaringUser: User;
   mandate: Mandate;
   comment: string;
+}
+
+export interface User {
+  id: number;
+  fullName: string;
+  isAdmin: boolean;
+  email: string;
+  organisations: Organisation[] | number[];
+  deliveries: Delivery[];
 }