+++ src/Geopilot.Frontend/src/Header.tsx
@@ -10,6 +10,7 @@ import {
   List,
   ListItem,
   ListItemButton,
+  ListItemIcon,
   ListItemText,
   Toolbar,
   Typography,
@@ -22,6 +23,7 @@ import { AdminTemplate } from "./auth/AdminTemplate.js";
 import { useLocation, useNavigate } from "react-router-dom";
 import { ClientSettings } from "./AppInterfaces";
 import { FC, useState } from "react";
+import OpenInNewIcon from "@mui/icons-material/OpenInNew";
 
 interface HeaderProps {
   clientSettings: ClientSettings;
@@ -38,6 +40,13 @@ export const Header: FC<HeaderProps> = ({ clientSettings, hasDrawerToggle, handl
 
   const toggleUserMenu = (newOpen: boolean) => () => {
     setUserMenuOpen(newOpen);
+  };
+
+  const isActive = (path: string) => {
+    if (path === "") {
+      return location.pathname === "/";
+    }
+    return location.pathname.split("/").includes(path);
   };
 
   return (
@@ -117,6 +126,7 @@ export const Header: FC<HeaderProps> = ({ clientSettings, hasDrawerToggle, handl
             <List>
               <ListItem key={t("delivery").toUpperCase()} disablePadding>
                 <ListItemButton
+                  selected={isActive("")}
                   onClick={() => {
                     navigate("/");
                   }}>
@@ -126,6 +136,7 @@ export const Header: FC<HeaderProps> = ({ clientSettings, hasDrawerToggle, handl
               <AdminTemplate>
                 <ListItem key={t("administration").toUpperCase()} disablePadding>
                   <ListItemButton
+                    selected={isActive("admin")}
                     onClick={() => {
                       navigate("/admin");
                     }}>
@@ -134,10 +145,14 @@ export const Header: FC<HeaderProps> = ({ clientSettings, hasDrawerToggle, handl
                 </ListItem>
                 <ListItem key={t("stacBrowser").toUpperCase()} disablePadding>
                   <ListItemButton
+                    selected={isActive("browser")}
                     onClick={() => {
-                      window.location.href = "/browser";
+                      window.open("/browser", "_blank");
                     }}>
                     <ListItemText primary={t("stacBrowser").toUpperCase()} />
+                    <ListItemIcon>
+                      <OpenInNewIcon fontSize="small" />
+                    </ListItemIcon>
                   </ListItemButton>
                 </ListItem>
               </AdminTemplate>