+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Add Cypress test support.
 - Added localization.
 - Added separate administration area and user navigation menu to switch between delivery, administration and STAC browser.
+- Added grid to manage mandates in administration area.
+- Added grid to manage organisations in administration area.
+- Added grid to manage users in administration area.
 - Added local Keycloak server for development.
 
 ### Changed
@@ -15,6 +18,7 @@
 - Rename _Abgabe_ to _Lieferung_.
 - Refactored delivery overview to use only [MUI](https://mui.com/material-ui/) components.
 - Use Typescript for new components.
+- STAC browser now opens in a new tab.
 - Use react-oidc-context for authentication.
 
 ## v1.0.93 - 2024-05-14